refactor: :memo: renames poo rejected file name to match component name
